Fixed the comments after review
